Add Zod schema and validation middleware update
