+++ back/src/middlewares/zod/validateCreateNewAdmin.js
@@ -1,0 +1,32 @@
+import { ZodError } from "zod";
+import { createAdminSchema } from "../../zodSchema/adminCreationValidationSchema.js";
+
+/********************************************* 
+ * Middleware de validation d'une inscription a la newsletter
+ *********************************************/
+
+export function validateCreateNewAdmin() {
+    return(req, res, next) => {
+        try{
+            //validation des données
+            const validateData = createAdminSchema.parse(req.body);
+            //injecte les données validées dans req.body
+            req.body = validateData;
+            //passer au prochain middleware
+            next();
+        } catch (error) {
+            if (error instanceof ZodError) {
+                return res.status(400).json({
+                    success: false,
+                    message: "Validation error",
+                    errors: error.errors.map(err=> ({
+                        field: err.path.join("."),
+                        message: err.message
+                    }))
+                });       
+            }
+            //autres erreur
+            next(error);
+        }
+    };
+}
